using cleanText.py to update 'त्स' -> 'थ्स') & 'प्स' -> 'फ्स' across all files
Excluding shlokas and nyAsa in prANAyAma

+++ cleanText.py
@@ -3,8 +3,9 @@ import sys
 
 
 def main():
-    repl_list = [# ('त्स्', 'थ्स्'), ('प्स्', 'फ्स्'),
-                 # ('र्द्ध्', 'र्ध्'),
+    repl_list = [# ('र्द्ध्', 'र्ध्'), ('र्द्ध', 'र्ध'),
+                 ('त्स्', 'थ्स्'), ('प्स्', 'फ्स्'),
+                 ('त्स', 'थ्स'), ('प्स', 'फ्स'),
                  ('', '᳚'),
                  ('॑', 'ः॑'), ('॒', 'ः॒'), ('᳚', 'ः᳚'),
                  ('', 'ꣳ'), ('', 'ꣳ॑'), ('', 'ꣳ᳚'), ('', 'ꣳ॒'),
@@ -14,12 +15,17 @@ def main():
         with open(file, 'r') as in_f:
             file_lines = in_f.readlines()
 
+        noReplaces = True
         for old, new in repl_list:
             for i in range(len(file_lines)):
-                file_lines[i] = file_lines[i].replace(old, new)
+                file_lines_new = file_lines[i].replace(old, new)
+                if file_lines_new != file_lines[i]:
+                    file_lines[i] = file_lines_new
+                    noReplaces = False
 
-        with open(file, 'w') as out_f:
-            out_f.writelines(file_lines)
+        if not noReplaces:
+            with open(file, 'w') as out_f:
+                out_f.writelines(file_lines)
 
 
 if __name__ == '__main__':
